voice/requirements.txt
# Mac (discord bot + client)
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0
httpx>=0.25.0
aiohttp>=3.9.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
silero-vad>=4.0.0
onnxruntime>=1.16.0
torch>=2.1.0
numpy>=1.24.0

# PC fisso (FastAPI server — aggiungi qui sotto)
# fastapi>=0.104.0
# uvicorn[standard]>=0.24.0
# faster-whisper>=1.0.0
# torch>=2.1.0  (con CUDA)
# torchaudio>=2.1.0
# elevenlabs>=1.0.0
